test: Cleanup the integration test suite
